Rename util module to dnsdiag (#123)
util is too generic to be used as shared module.

Fixes #122
