+++ dnsdiag/shared.py
